+++ scripts/make/build-docker.sh
@@ -87,7 +87,7 @@ readonly docker_image_full_name docker_tags
 
 # Copy the binaries into a new directory under new names, so that it's easier to
 # COPY them later.  DO NOT remove the trailing underscores.  See file
-# scripts/make/Dockerfile.
+# docker/Dockerfile.
 dist_docker="${dist_dir}/docker"
 readonly dist_docker
 
@@ -105,6 +105,18 @@ cp "${dist_dir}/AdGuardHome_linux_arm_7/AdGuardHome/AdGuardHome"\
 cp "${dist_dir}/AdGuardHome_linux_ppc64le/AdGuardHome/AdGuardHome"\
 	"${dist_docker}/AdGuardHome_linux_ppc64le_"
 
+# Copy the helper scripts.  See file docker/Dockerfile.
+dist_docker_scripts="${dist_docker}/scripts"
+readonly dist_docker_scripts
+
+mkdir -p "$dist_docker_scripts"
+cp "./docker/dns-bind.awk"\
+	"${dist_docker_scripts}/dns-bind.awk"
+cp "./docker/web-bind.awk"\
+	"${dist_docker_scripts}/web-bind.awk"
+cp "./docker/healthcheck.sh"\
+	"${dist_docker_scripts}/healthcheck.sh"
+
 # Don't use quotes with $docker_tags and $debug_flags because we want word
 # splitting and or an empty space if tags are empty.
 $sudo_cmd docker\
@@ -118,5 +130,5 @@ $sudo_cmd docker\
 	--platform "$docker_platforms"\
 	$docker_tags\
 	-t "$docker_image_full_name"\
-	-f ./scripts/make/Dockerfile\
+	-f ./docker/Dockerfile\
 	.